requirements.txt
click==7.1.2
databases==0.3.2
fastapi==0.58.0
h11==0.9.0
httptools==0.1.1
motor==2.1.0
pydantic==1.5.1
pymongo==3.10.1
python-dotenv==0.13.0
SQLAlchemy==1.3.17
starlette==0.13.4
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
